lastback/backend/requirements.txt
absl-py==0.12.0
asgiref==3.3.4
astroid==2.5.6
astunparse==1.6.3
backcall==0.2.0
CacheControl==0.12.6
cached-property==1.5.2
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==8.0.1
colorama==0.4.4
cycler==0.10.0
dataclasses==0.6
decorator==5.0.9
deepface==0.0.53
Django==3.2.3
django-appconf==1.0.4
django-cors-headers==3.7.0
django-imagekit==4.0.2
djangorestframework==3.12.4
djongo==1.3.6
filelock==3.0.12
firebase-admin==5.0.0
Flask==2.0.1
flatbuffers==1.12
gast==0.4.0
gdown==3.13.0
google-api-core==1.28.0
google-api-python-client==2.7.0
google-auth==1.30.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
google-cloud-core==1.6.0
google-cloud-firestore==2.1.1
google-cloud-storage==1.38.0
google-crc32c==1.1.2
google-pasta==0.2.0
google-resumable-media==1.3.0
googleapis-common-protos==1.53.0
grpcio==1.34.1
gunicorn==20.1.0
h5py==3.1.0
httplib2==0.19.1
idna==2.10
image==1.5.33
importlib-metadata==4.4.0
ipython==7.16.1
ipython-genutils==0.2.0
isort==5.8.0
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
Keras==2.4.3
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
lazy-object-proxy==1.6.0
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.3.4
mccabe==0.6.1
msgpack==1.0.2
mtcnn==0.1.0
mysqlclient==2.0.3
numpy==1.19.5
oauthlib==3.1.1
opencv-python==4.5.1.48
opt-einsum==3.3.0
packaging==20.9
paho-mqtt==1.5.1
pandas==1.1.5
parso==0.8.2
pickleshare==0.7.5
pilkit==2.0
Pillow==8.2.0
prompt-toolkit==3.0.18
proto-plus==1.18.1
protobuf==3.17.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.9.0
pylint==2.8.3
pymongo==3.11.4
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
retina-face==0.0.4
rsa==4.7.2
scipy==1.5.4
seaborn==0.11.1
six==1.15.0
sqlparse==0.2.4
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
termcolor==1.1.0
toml==0.10.2
torch==1.8.1
torchvision==0.9.1
tqdm==4.61.0
traitlets==4.3.3
typed-ast==1.4.3
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.5
wcwidth==0.2.5
Werkzeug==2.0.1
wincertstore==0.2
wrapt==1.12.1
zipp==3.4.1
